Add : User register verify route, activeUserAccount controller with jwt token verification

+++ server/controllers/user.controller.js
@@ -1,12 +1,12 @@
 import { errorHandler, successHandler } from "./responseHandler.controller.js";
 import User from "../models/user.model.js";
 import createError from "http-errors";
-import mongoose from "mongoose";
 import { findWithId } from "../services/findWithId.js";
 import deleteImage from "../helper/deleteImage.js";
 import { generateToken } from "../helper/jsonwebtoken.js";
 import { clientUrl, jwtActivationKey } from "../src/secret.js";
 import sendEmailWithNodeMailer from "../helper/email.js";
+import jwt from "jsonwebtoken";
 
 const getUsers = async (req, res, next) => {
   try {
@@ -58,7 +58,7 @@ const getUserById = async (req, res, next) => {
     const id = req.params.id;
     const options = { password: 0 };
     const user = await findWithId(User, id, options);
-      
+
     if (!user) {
       return errorHandler(res, { statusCode: 404, message: "User not found" });
     } else {
@@ -66,7 +66,7 @@ const getUserById = async (req, res, next) => {
         statusCode: 200,
         message: "User found successfully",
         payload: {
-          user
+          user,
         },
       });
     }
@@ -85,14 +85,12 @@ const deleteUserById = async (req, res, next) => {
 
     deleteImage(userImagePath);
 
-    await User.findByIdAndDelete({ _id: id ,
-      isAdmin:false
-    });
+    await User.findByIdAndDelete({ _id: id, isAdmin: false });
 
-      return successHandler(res, {
-        statusCode: 200,
-        message: "User deleted successfully",
-      });
+    return successHandler(res, {
+      statusCode: 200,
+      message: "User deleted successfully",
+    });
   } catch (error) {
     next(error);
   }
@@ -100,19 +98,19 @@ const deleteUserById = async (req, res, next) => {
 
 const processRegister = async (req, res, next) => {
   try {
-    const {name, email, password, phone, address } = req.body;
+    const { name, email, password, phone, address } = req.body;
 
-    const user = new User({
+    const user = {
       name,
       email,
       password,
       phone,
       address,
-    });
+    };
 
-    const userExists = await User.findOne({email:email});
-    if(userExists){
-      throw createError(409,"User already exists. Please login");
+    const userExists = await User.findOne({ email: email });
+    if (userExists) {
+      throw createError(409, "User already exists. Please login");
     }
 
     const token = generateToken(
@@ -134,17 +132,16 @@ const processRegister = async (req, res, next) => {
       await sendEmailWithNodeMailer(emailData);
       console.log("Email sent");
     } catch (error) {
-      
       throw createError(500, "Email not sent");
     }
 
     return successHandler(res, {
       statusCode: 200,
       message: `Please check your ${emailData.email} to activate your account`,
-      payload:{
+      payload: {
         user,
-        token
-      }
+        token,
+      },
     });
   } catch (error) {
     console.log(error);
@@ -152,5 +149,46 @@ const processRegister = async (req, res, next) => {
   }
 };
 
+const activeUserAccount = async (req, res, next) => {
+  try {
+    const token = req.body.token;
 
-export { getUsers, getUserById, deleteUserById , processRegister};
+    if (!token) throw createError(400, "Token not found");
+
+    try {
+      const decoded = jwt.verify(token, jwtActivationKey);
+      if (!decoded) throw createError(400, "Unable to verify user");
+      console.log(decoded);
+
+      const userExists = await User.findOne({ email: decoded.email });
+      if(userExists){
+        throw createError(409, "User already exists. Please login");
+      }
+
+      await User.create(decoded);
+
+      return successHandler(res, {
+        statusCode: 201,
+        message: "User account registered successfully",
+      });
+    } catch (error) {
+      if (error.name === "TokenExpiredError") {
+        throw createError(401, "Token has Expired");
+      } else if (error.name === "JsonWebTokenError") {
+        throw createError(401, "Invalid Token");
+      } else {
+        throw createError(400, "Something went wrong. Unable to verify user. Please register again");
+      }
+    }
+  } catch (error) {
+    next(error);
+  }
+};
+
+export {
+  getUsers,
+  getUserById,
+  deleteUserById,
+  processRegister,
+  activeUserAccount,
+};
